+++ test/InvariantTests.t.sol
@@ -36,14 +36,14 @@ contract InvariantTests is StdInvariant, Test {
     /// @notice Total token supply should always equal sum of individual balances
     function invariant_TokenSupplyConsistency() public view {
         uint256[] memory allTokenIds = handler.getAllTokenIds();
-        
-        for (uint i = 0; i < allTokenIds.length; i++) {
+
+        for (uint256 i = 0; i < allTokenIds.length; i++) {
             uint256 tokenId = allTokenIds[i];
             uint256 totalSupply = assemble.totalSupply(tokenId);
             uint256 sumOfBalances = 0;
 
             address[] memory holders = handler.getTokenHolders(tokenId);
-            for (uint j = 0; j < holders.length; j++) {
+            for (uint256 j = 0; j < holders.length; j++) {
                 sumOfBalances += assemble.balanceOf(holders[j], tokenId);
             }
 
@@ -54,16 +54,16 @@ contract InvariantTests is StdInvariant, Test {
     /// @notice Payment splits for all events must always total 100%
     function invariant_PaymentSplitsTotal100Percent() public view {
         uint256[] memory allEventIds = handler.getAllEventIds();
-        
-        for (uint i = 0; i < allEventIds.length; i++) {
+
+        for (uint256 i = 0; i < allEventIds.length; i++) {
             uint256 eventId = allEventIds[i];
             Assemble.PaymentSplit[] memory splits = assemble.getPaymentSplits(eventId);
-            
+
             uint256 totalBps = 0;
-            for (uint j = 0; j < splits.length; j++) {
+            for (uint256 j = 0; j < splits.length; j++) {
                 totalBps += splits[j].basisPoints;
             }
-            
+
             assertEq(totalBps, 10_000, "Payment splits must always total 100%");
         }
     }
@@ -71,13 +71,12 @@ contract InvariantTests is StdInvariant, Test {
     /// @notice Soulbound tokens should never be transferable
     function invariant_SoulboundTokensNotTransferable() public view {
         uint256[] memory allTokenIds = handler.getAllTokenIds();
-        
-        for (uint i = 0; i < allTokenIds.length; i++) {
+
+        for (uint256 i = 0; i < allTokenIds.length; i++) {
             uint256 tokenId = allTokenIds[i];
             Assemble.TokenType tokenType = Assemble.TokenType(tokenId >> 248);
-            
-            if (tokenType == Assemble.TokenType.ATTENDANCE_BADGE || 
-                tokenType == Assemble.TokenType.ORGANIZER_CRED) {
+
+            if (tokenType == Assemble.TokenType.ATTENDANCE_BADGE || tokenType == Assemble.TokenType.ORGANIZER_CRED) {
                 // These should be soulbound - test by checking they can't be transferred
                 // This is tested in the actual transfer function, not here
                 assertTrue(true, "Soulbound invariant checked in transfer function");
@@ -89,20 +88,20 @@ contract InvariantTests is StdInvariant, Test {
     function invariant_RefundsNeverExceedPayments() public view {
         uint256[] memory allEventIds = handler.getAllEventIds();
         address[] memory allUsers = handler.getAllUsers();
-        
-        for (uint i = 0; i < allEventIds.length; i++) {
+
+        for (uint256 i = 0; i < allEventIds.length; i++) {
             uint256 eventId = allEventIds[i];
-            
+
             if (assemble.isEventCancelled(eventId)) {
-                for (uint j = 0; j < allUsers.length; j++) {
+                for (uint256 j = 0; j < allUsers.length; j++) {
                     address user = allUsers[j];
                     (uint256 ticketRefund, uint256 tipRefund) = assemble.getRefundAmounts(eventId, user);
-                    
+
                     // In our current implementation, we track exactly what users paid
                     // So refunds should equal payments (not exceed)
                     uint256 totalRefund = ticketRefund + tipRefund;
                     uint256 maxPossiblePayment = handler.getMaxPaymentForUser(eventId, user);
-                    
+
                     assertLe(totalRefund, maxPossiblePayment, "Refunds should not exceed payments");
                 }
             }
@@ -112,29 +111,24 @@ contract InvariantTests is StdInvariant, Test {
     /// @notice Event capacity should never be exceeded
     function invariant_EventCapacityNotExceeded() public view {
         uint256[] memory allEventIds = handler.getAllEventIds();
-        
-        for (uint i = 0; i < allEventIds.length; i++) {
+
+        for (uint256 i = 0; i < allEventIds.length; i++) {
             uint256 eventId = allEventIds[i];
-            (, , uint32 capacity, , ,) = assemble.events(eventId);
-            
+            (,, uint32 capacity,,,) = assemble.events(eventId);
+
             // Sum all ticket tiers sold
             uint256 totalSold = 0;
-            for (uint256 tierId = 0; tierId < 10; tierId++) { // Max reasonable tiers
+            for (uint256 tierId = 0; tierId < 10; tierId++) {
+                // Max reasonable tiers
                 try assemble.ticketTiers(eventId, tierId) returns (
-                    string memory,
-                    uint256,
-                    uint256,
-                    uint256 sold,
-                    uint256,
-                    uint256,
-                    bool
+                    string memory, uint256, uint256, uint256 sold, uint256, uint256, bool
                 ) {
                     totalSold += sold;
                 } catch {
                     break; // No more tiers
                 }
             }
-            
+
             assertLe(totalSold, capacity, "Total tickets sold should not exceed capacity");
         }
     }
@@ -143,7 +137,7 @@ contract InvariantTests is StdInvariant, Test {
     function invariant_ProtocolFeesCorrect() public view {
         uint256 protocolFeeBps = assemble.protocolFeeBps();
         assertLe(protocolFeeBps, 1000, "Protocol fee should not exceed 10%");
-        
+
         // Protocol fee recipient should have accumulated fees
         uint256 protocolFees = assemble.pendingWithdrawals(feeTo);
         assertGe(protocolFees, 0, "Protocol fees should be non-negative");
@@ -152,12 +146,12 @@ contract InvariantTests is StdInvariant, Test {
     /// @notice Friend relationships should be consistent
     function invariant_FriendRelationshipsConsistent() public view {
         address[] memory allUsers = handler.getAllUsers();
-        
-        for (uint i = 0; i < allUsers.length; i++) {
+
+        for (uint256 i = 0; i < allUsers.length; i++) {
             address user = allUsers[i];
             address[] memory friends = assemble.getFriends(user);
-            
-            for (uint j = 0; j < friends.length; j++) {
+
+            for (uint256 j = 0; j < friends.length; j++) {
                 address friend = friends[j];
                 assertTrue(assemble.isFriend(user, friend), "Friend list should be consistent with mapping");
                 assertNotEq(user, friend, "User should not be friends with themselves");
@@ -169,15 +163,16 @@ contract InvariantTests is StdInvariant, Test {
     /// @notice Comment integrity should be maintained
     function invariant_CommentIntegrity() public view {
         uint256 nextCommentId = assemble.nextCommentId();
-        
+
         for (uint256 commentId = 1; commentId < nextCommentId; commentId++) {
             CommentLibrary.Comment memory comment = assemble.getComment(commentId);
-            
-            if (comment.timestamp > 0) { // Comment exists
+
+            if (comment.timestamp > 0) {
+                // Comment exists
                 assertNotEq(comment.author, address(0), "Comment author should not be zero");
                 assertGe(comment.likes, 0, "Likes should be non-negative");
                 assertLe(bytes(comment.content).length, 1000, "Comment should not exceed max length");
-                
+
                 // If comment has parent, parent should exist
                 if (comment.parentId > 0) {
                     CommentLibrary.Comment memory parentComment = assemble.getComment(comment.parentId);
@@ -192,14 +187,14 @@ contract InvariantTests is StdInvariant, Test {
 /// @notice Provides controlled randomized actions for the protocol
 contract AssembleHandler is Test {
     Assemble public assemble;
-    
+
     // Track state for invariants
     uint256[] public allEventIds;
     uint256[] public allTokenIds;
     address[] public allUsers;
     mapping(uint256 => address[]) public tokenHolders;
     mapping(uint256 => mapping(address => uint256)) public maxUserPayments;
-    
+
     // Ghost variables for tracking
     uint256 public totalProtocolFees;
     uint256 public totalRefundsClaimed;
@@ -215,9 +210,9 @@ contract AssembleHandler is Test {
 
     constructor(Assemble _assemble) {
         assemble = _assemble;
-        
+
         // Initialize some users
-        for (uint i = 0; i < 10; i++) {
+        for (uint256 i = 0; i < 10; i++) {
             address user = makeAddr(string(abi.encodePacked("user", vm.toString(i))));
             allUsers.push(user);
             vm.deal(user, 1000 ether);
@@ -229,7 +224,10 @@ contract AssembleHandler is Test {
         uint256 startTimeOffset,
         uint256 capacity,
         uint256 tierPrice
-    ) public useActor(actorSeed) {
+    )
+        public
+        useActor(actorSeed)
+    {
         startTimeOffset = bound(startTimeOffset, 1 hours, 30 days);
         capacity = bound(capacity, 1, 1000);
         tierPrice = bound(tierPrice, 0, 1 ether);
@@ -263,38 +261,35 @@ contract AssembleHandler is Test {
         allEventIds.push(eventId);
     }
 
-    function purchaseTickets(
-        uint256 actorSeed,
-        uint256 eventIndex,
-        uint256 quantity
-    ) public useActor(actorSeed) {
+    function purchaseTickets(uint256 actorSeed, uint256 eventIndex, uint256 quantity) public useActor(actorSeed) {
         if (allEventIds.length == 0) return;
-        
+
         eventIndex = bound(eventIndex, 0, allEventIds.length - 1);
         quantity = bound(quantity, 1, 10);
-        
+
         uint256 eventId = allEventIds[eventIndex];
         uint256 price = assemble.calculatePrice(eventId, 0, quantity);
-        
+
         if (price > 0 && msg.sender.balance >= price) {
-            try assemble.purchaseTickets{value: price}(eventId, 0, quantity) {
+            try assemble.purchaseTickets{ value: price }(eventId, 0, quantity) {
                 // Track multiple token IDs (one for each ticket purchased)
-                (, , , uint256 soldBefore, , ,) = assemble.ticketTiers(eventId, 0);
-                
+                (,,, uint256 soldBefore,,,) = assemble.ticketTiers(eventId, 0);
+
                 for (uint256 i = 0; i < quantity; i++) {
                     uint256 serialNumber = soldBefore - quantity + i + 1;
-                    uint256 tokenId = assemble.generateTokenId(Assemble.TokenType.EVENT_TICKET, eventId, 0, serialNumber);
-                    
+                    uint256 tokenId =
+                        assemble.generateTokenId(Assemble.TokenType.EVENT_TICKET, eventId, 0, serialNumber);
+
                     if (!_contains(allTokenIds, tokenId)) {
                         allTokenIds.push(tokenId);
                     }
-                    
+
                     // Track token holders
                     if (!_contains(tokenHolders[tokenId], msg.sender)) {
                         tokenHolders[tokenId].push(msg.sender);
                     }
                 }
-                
+
                 // Track max payments
                 maxUserPayments[eventId][msg.sender] += price;
             } catch {
@@ -305,25 +300,21 @@ contract AssembleHandler is Test {
 
     function addFriend(uint256 actorSeed, uint256 friendIndex) public useActor(actorSeed) {
         if (allUsers.length < 2) return;
-        
+
         friendIndex = bound(friendIndex, 0, allUsers.length - 1);
         address friend = allUsers[friendIndex];
-        
+
         if (friend != msg.sender && !assemble.isFriend(msg.sender, friend)) {
             assemble.addFriend(friend);
         }
     }
 
-    function postComment(
-        uint256 actorSeed,
-        uint256 eventIndex,
-        string calldata content
-    ) public useActor(actorSeed) {
+    function postComment(uint256 actorSeed, uint256 eventIndex, string calldata content) public useActor(actorSeed) {
         if (allEventIds.length == 0) return;
-        
+
         eventIndex = bound(eventIndex, 0, allEventIds.length - 1);
         uint256 eventId = allEventIds[eventIndex];
-        
+
         if (bytes(content).length > 0 && bytes(content).length <= 1000) {
             try assemble.postComment(eventId, content, 0) {
                 // Comment posted successfully
@@ -333,20 +324,16 @@ contract AssembleHandler is Test {
         }
     }
 
-    function tipEvent(
-        uint256 actorSeed,
-        uint256 eventIndex,
-        uint256 tipAmount
-    ) public useActor(actorSeed) {
+    function tipEvent(uint256 actorSeed, uint256 eventIndex, uint256 tipAmount) public useActor(actorSeed) {
         if (allEventIds.length == 0) return;
-        
+
         eventIndex = bound(eventIndex, 0, allEventIds.length - 1);
         tipAmount = bound(tipAmount, 0.001 ether, 0.1 ether);
-        
+
         uint256 eventId = allEventIds[eventIndex];
-        
+
         if (msg.sender.balance >= tipAmount) {
-            try assemble.tipEvent{value: tipAmount}(eventId) {
+            try assemble.tipEvent{ value: tipAmount }(eventId) {
                 maxUserPayments[eventId][msg.sender] += tipAmount;
             } catch {
                 // Tip failed, continue
@@ -356,10 +343,10 @@ contract AssembleHandler is Test {
 
     function cancelEvent(uint256 actorSeed, uint256 eventIndex) public useActor(actorSeed) {
         if (allEventIds.length == 0) return;
-        
+
         eventIndex = bound(eventIndex, 0, allEventIds.length - 1);
         uint256 eventId = allEventIds[eventIndex];
-        
+
         if (assemble.eventOrganizers(eventId) == msg.sender) {
             try assemble.cancelEvent(eventId) {
                 // Event cancelled
@@ -381,10 +368,10 @@ contract AssembleHandler is Test {
 
     function claimTicketRefund(uint256 actorSeed, uint256 eventIndex) public useActor(actorSeed) {
         if (allEventIds.length == 0) return;
-        
+
         eventIndex = bound(eventIndex, 0, allEventIds.length - 1);
         uint256 eventId = allEventIds[eventIndex];
-        
+
         if (assemble.isEventCancelled(eventId)) {
             (uint256 refundAmount,) = assemble.getRefundAmounts(eventId, msg.sender);
             if (refundAmount > 0) {
@@ -399,10 +386,10 @@ contract AssembleHandler is Test {
 
     function claimTipRefund(uint256 actorSeed, uint256 eventIndex) public useActor(actorSeed) {
         if (allEventIds.length == 0) return;
-        
+
         eventIndex = bound(eventIndex, 0, allEventIds.length - 1);
         uint256 eventId = allEventIds[eventIndex];
-        
+
         if (assemble.isEventCancelled(eventId)) {
             (, uint256 refundAmount) = assemble.getRefundAmounts(eventId, msg.sender);
             if (refundAmount > 0) {
@@ -438,16 +425,16 @@ contract AssembleHandler is Test {
 
     // Helper functions
     function _contains(uint256[] memory array, uint256 value) internal pure returns (bool) {
-        for (uint i = 0; i < array.length; i++) {
+        for (uint256 i = 0; i < array.length; i++) {
             if (array[i] == value) return true;
         }
         return false;
     }
 
     function _contains(address[] memory array, address value) internal pure returns (bool) {
-        for (uint i = 0; i < array.length; i++) {
+        for (uint256 i = 0; i < array.length; i++) {
             if (array[i] == value) return true;
         }
         return false;
     }
-} 
+}
